Remove types associated only with v1 api
