+++ packages/common/src/pub/util.ts
@@ -1,7 +1,2 @@
 // Utility functions needed in web/nodejs packages
-export {
-    buildJsonResponseReviver,
-    intToStringTransformer,
-    isValidHash,
-    stringToInt,
-} from '../util.js';
+export { isValidHash, stringToInt } from '../util.js';
