+++ packages/common/src/types/rejectReason.ts
@@ -155,14 +155,6 @@ export interface BakerIdRejectReason {
     contents: BakerId;
 }
 
-/**
- * @deprecated This is type for describing return types for the JSON-RPC client and the V1 gRPC client, both of which have been deprecated
- */
-export interface NumberRejectReason {
-    tag: BakerIdRejectReasonTag;
-    contents: number;
-}
-
 export interface SimpleRejectReason {
     tag: SimpleRejectReasonTag;
 }
@@ -175,14 +167,6 @@ export interface InvalidReceiveMethod {
     };
 }
 
-/**
- * @deprecated This is type for describing return types for the JSON-RPC client and the V1 gRPC client, both of which have been deprecated
- */
-export interface InvalidReceiveMethodV1 {
-    tag: RejectReasonTag.InvalidReceiveMethod;
-    contents: [HexString, string]; // [moduleRef, receiveName]
-}
-
 export interface InvalidInitMethod {
     tag: RejectReasonTag.InvalidInitMethod;
     contents: {
@@ -191,28 +175,12 @@ export interface InvalidInitMethod {
     };
 }
 
-/**
- * @deprecated This is type for describing return types for the JSON-RPC client and the V1 gRPC client, both of which have been deprecated
- */
-export interface InvalidInitMethodV1 {
-    tag: RejectReasonTag.InvalidInitMethod;
-    contents: [HexString, string]; // [moduleRef, initName]
-}
-
 export interface AmountTooLarge {
     tag: RejectReasonTag.AmountTooLarge;
     contents: {
         address: Address;
         amount: Amount;
     };
-}
-
-/**
- * @deprecated This is type for describing return types for the JSON-RPC client and the V1 gRPC client, both of which have been deprecated
- */
-export interface AmountTooLargeV1 {
-    tag: RejectReasonTag.AmountTooLarge;
-    contents: [Address, DigitString]; // [address, amount]
 }
 
 export interface InvalidContractAddress {
@@ -243,13 +211,3 @@ export type RejectReason =
     | InvalidReceiveMethod
     | InvalidInitMethod
     | AmountTooLarge;
-
-/**
- * @deprecated This is type for describing return types for the JSON-RPC client and the V1 gRPC client, both of which have been deprecated
- */
-export type RejectReasonV1 =
-    | RejectReasonCommon
-    | NumberRejectReason
-    | InvalidReceiveMethodV1
-    | InvalidInitMethodV1
-    | AmountTooLargeV1;
